rename actions to adminapi
