--- app/domain/actions/__init__.py
@@ -1,23 +1,0 @@
-from app.domain.actions.add_data import add_data
-from app.domain.actions.create_source import create_source
-from app.domain.actions.create_table import create_table
-from app.domain.actions.get_source import get_source
-from app.domain.actions.get_task_info import get_task_info
-from app.domain.actions.login import login
-from app.domain.actions.set_table_status import set_table_status
-from app.domain.actions.start_task import start_task
-from app.domain.actions.table_process import table_process
-from app.domain.actions.table_status_stats import table_status_stats
-
-__all__ = [
-    "add_data",
-    "create_source",
-    "create_table",
-    "get_source",
-    "get_task_info",
-    "login",
-    "start_task",
-    "table_process",
-    "table_status_stats",
-    "set_table_status",
-]
