+++ app/domain/adminapi/__init__.py
@@ -1,0 +1,23 @@
+from app.domain.adminapi.add_data import add_data
+from app.domain.adminapi.create_source import create_source
+from app.domain.adminapi.create_table import create_table
+from app.domain.adminapi.get_source import get_source
+from app.domain.adminapi.get_task_info import get_task_info
+from app.domain.adminapi.login import login
+from app.domain.adminapi.set_table_status import set_table_status
+from app.domain.adminapi.start_task import start_task
+from app.domain.adminapi.table_process import table_process
+from app.domain.adminapi.table_status_stats import table_status_stats
+
+__all__ = [
+    "add_data",
+    "create_source",
+    "create_table",
+    "get_source",
+    "get_task_info",
+    "login",
+    "start_task",
+    "table_process",
+    "table_status_stats",
+    "set_table_status",
+]
