+++ app/domain/adminapi/table_status_stats.py
