+++ app/presentation/adminapi/create_source.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load, validate
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -53,7 +53,7 @@ async def create_source_handler(dpt: depot.Depot, r: web.Request) -> responses.A
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.create_source(dpt, request))
+    return responses.APIOkResponse(adminapi.create_source(dpt, request))
 
 
 description = common.handler_description(
