+++ app/presentation/adminapi/get_source.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -50,7 +50,7 @@ async def get_source_handler(dpt: depot.Depot, r: web.Request) -> responses.APIO
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.get_source(dpt, request))
+    return responses.APIOkResponse(adminapi.get_source(dpt, request))
 
 
 description = common.handler_description(
