+++ app/presentation/adminapi/get_task_info.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -52,7 +52,7 @@ async def get_task_info_handler(dpt: depot.Depot, r: web.Request) -> responses.A
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.get_task_info(dpt, request))
+    return responses.APIOkResponse(adminapi.get_task_info(dpt, request))
 
 
 description = common.handler_description(
