+++ app/presentation/adminapi/login.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -49,7 +49,7 @@ async def login_handler(dpt: depot.Depot, r: web.Request) -> responses.APIOkResp
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.login(dpt, request))
+    return responses.APIOkResponse(adminapi.login(dpt, request))
 
 
 description = common.handler_description(
