+++ app/presentation/adminapi/set_table_status.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -59,7 +59,7 @@ async def set_table_status_handler(dpt: depot.Depot, r: web.Request) -> response
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.set_table_status(dpt, request))
+    return responses.APIOkResponse(adminapi.set_table_status(dpt, request))
 
 
 description = common.handler_description(
