+++ app/presentation/adminapi/start_task.py
@@ -3,7 +3,7 @@ from marshmallow import Schema, ValidationError, fields, post_load
 
 from app import schema
 from app.commands.adminapi import depot
-from app.domain import actions
+from app.domain import adminapi
 from app.lib.web import responses, server
 from app.lib.web.errors import RuleValidationError
 from app.presentation.adminapi import common
@@ -53,7 +53,7 @@ async def start_task_handler(dpt: depot.Depot, r: web.Request) -> responses.APIO
     except ValidationError as e:
         raise RuleValidationError(str(e)) from e
 
-    return responses.APIOkResponse(actions.start_task(dpt, request))
+    return responses.APIOkResponse(adminapi.start_task(dpt, request))
 
 
 description = common.handler_description(
