+++ tests/unit/domain/create_table_test.py
@@ -7,8 +7,8 @@ from parameterized import param, parameterized
 from app import entities, schema
 from app.commands.adminapi import depot
 from app.data import repositories
-from app.domain import actions
-from app.domain.actions.create_table import domain_descriptions_to_data, get_source_id
+from app.domain import adminapi
+from app.domain.adminapi.create_table import domain_descriptions_to_data, get_source_id
 from app.lib import testing
 from app.lib.storage import mapping
 from app.lib.web import errors
@@ -203,14 +203,14 @@ class CreateTableTest(unittest.TestCase):
 
         if err_substr is not None:
             with self.assertRaises(errors.RuleValidationError) as err:
-                _, _ = actions.create_table(
+                _, _ = adminapi.create_table(
                     self.depot,
                     request,
                 )
 
             self.assertIn(err_substr, err.exception.message())
         else:
-            resp, created = actions.create_table(
+            resp, created = adminapi.create_table(
                 self.depot,
                 request,
             )
