+++ tests/unit/domain/cross_identify_test.py
@@ -5,7 +5,7 @@ import astropy.units as u
 from astropy.coordinates import ICRS, angular_separation
 
 from app import entities
-from app.domain.actions.table_process import DEFAULT_INNER_RADIUS, cross_identification
+from app.domain.adminapi.table_process import DEFAULT_INNER_RADIUS, cross_identification
 from app.domain.cross_id_simultaneous_data_provider import SimpleSimultaneousDataProvider
 from app.domain.model.layer2 import Layer2Model
 from app.domain.model.params.cross_identification_result import (
